sony: kitakami: Fix iddd service
iddd service uses /idd partition which is ...by-name/diag block device.
We are using apps_log for /misc which is required for recovery
and it breaks /rca requirements

So this patch creates the symlink /idd/rca -> /rca
and fix statup-logger as following:

12-19 16:40:51.068   479   479 I Startup-Logger: Successfully read miscta unit value
12-19 16:40:51.068   479   479 I Startup-Logger: Entering should_remove_logs
12-19 16:40:51.068   479   479 I Startup-Logger: input value =0x50
12-19 16:40:51.068   479   479 I Startup-Logger: Removing /mnt/rca/startup-logger/ Done

Now iddd service is ready to serve the world.
12-19 15:32:44.832   435   435 I iddd    : Ready to serve the world.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: If5c1f63f89eeb2e4920ef41c289dc69f8c02ba03